requirements.txt
app==0.0.1
astroid==2.4.2
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
cachetools==5.3.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==8.0.1
cloudscraper==1.2.58
colorgram.py==1.2.0
cryptography==3.4.7
cssselect==1.1.0
dicttoxml==1.7.4
docutils==0.17.1
encrypter==0.1
fernet==1.0.1
Flask==2.0.1
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
html2text==2020.1.16
httplib2==0.21.0
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
json2xml==3.6.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.0.1
mccabe==0.6.1
Menu==3.2.2
MyApplication==0.1.0
numpy==1.24.2
oauthlib==3.2.2
packaging==20.9
pandas==1.5.3
pid==3.0.4
Pillow==9.4.0
playsound==1.2.2
pluggy==0.13.1
prettytable==2.1.0
protobuf==4.22.0
psycopg2-binary==2.9.5
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.4
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
service==0.6.0
setproctitle==1.3.2
six==1.16.0
soupsieve==2.2.1
spotipy==2.18.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.5
utils==1.0.1
wcwidth==0.2.5
Werkzeug==2.0.1
woob==3.0
wrapt==1.12.1
xmltodict==0.11.0
zipp==3.4.1
